Will not be using multer-storage-cloudinary
I wrote the implementation for it, but to install the library, it
requires to go from cloudnary v2.4 to v1.9. I dont want to go down an
entire version, so I reconsidered

There are other ways to accomplish this upload:
- multer disk storage -> upload to cloudinary -> delete file
- multer memory storage -> upload to cloudinary -> remove from memory
- presigned URL
- streaming direcly the file upload from frontend through backend to
cloudinary, as the backend is receiving the file. This is what we wanted
to achieve with multer-storage-cloudinary. Can use streamifier instead.

-Memory storage is really bad, because memory is most valuable resource.
This does not scale at all
-Disk storage is not that great because of constant read writes to disk
being expensive and slow, also requires good managing of server disk

Only good options are presigned URL or directly streaming:
- I am deicidng to shift to presigned URLs, because if I think about the
steps of error checking, they are the same:
1. In 1st request, send form data and validate it, if form inputs are
wrong, then dont even create the presigned cloudinary URL and send back
validation error instead
2. If we get presigned URl, then frontend sends file to cloudinary
3. If success, then it sends another request to backend to write the
DB document that will contain a link to this file. If creating the DB
document fails, then delete the cloudinary file

All of the above error checking steps would be the same regardless of
whether we use presigned URLs or streamed the upload through the backend
to cloudinary. The only extra thing to worry about is:
if the backend can't be reached on the 2nd request, then we would end
up with a 'ghost file' in cloudinary, with no corresponding DB document.
To fix this, we will be tagging the file on the intial upload with the
presigned URL, and upon the backends 2nd request, we will be untagging
this tag. Then we will create a cron job that periodically scans our
cloudinary, and removes any files with this special tag. I guess the
only problem with this would be that when we run this job, we could
delete the files of a user that is in the middle of their upload request
cycle (they are waiting for backend to remove tag). But we could easily
overcome this by checking something like tag existance AND the file is
older than 24 hours