Simple working pre-signed url function
TODO
- maybe simplify
- check validation of fields on presign request (product title and
description). This way we can send a validation error on this 1st
request without signing url and without making the client upload a file
- create route to write the db document. also include same validation
here (hm, maybe I should check that this somehow comes from a recent
signature ????)
- Also from last commit:
The only extra thing to worry about is:
if the backend can't be reached on the 2nd request, then we would end
up with a 'ghost file' in cloudinary, with no corresponding DB document.
To fix this, we will be tagging the file on the intial upload with the
presigned URL, and upon the backends 2nd request, we will be untagging
this tag. Then we will create a cron job that periodically scans our
cloudinary, and removes any files with this special tag. I guess the
only problem with this would be that when we run this job, we could
delete the files of a user that is in the middle of their upload request
cycle (they are waiting for backend to remove tag). But we could easily
overcome this by checking something like tag existance AND the file is
older than 24 hours