Simplified and added simple skeleton for uploading
TODO
Importantly though, right now I am doing things slightly wrong on the
route that stores the document into DB:
We need to store the image URL into DB, but we are only acceptin imageId
(cloudinary public Id) as req param. We cant accept 2 body poarams
imageID and imageURL, because then when we save the document, we'd be
trusting that the imageURL corresponds to the imageID....
So instead, we should either switch to accepting a imageURL and
extracting the publicID from it, or we can switch to using
uploader.explicit and switching to using adding a context linked=false
instead.
I think just switch to above, because I can just foresee problems with
url input inconsitencies for example they send in
wrongUrl/correctPublicId.jpg, and I'd be extracting and
finding the image at correctPublicId, but then storing the wrongUrl in
database

TODO PREV (still aplpies)
- maybe simplify
- check validation of fields on presign request (product title and
description). This way we can send a validation error on this 1st
request without signing url and without making the client upload a file
- create route to write the db document. also include same validation
here (hm, maybe I should check that this somehow comes from a recent
signature ????)
- Also from last commit:
The only extra thing to worry about is:
if the backend can't be reached on the 2nd request, then we would end
up with a 'ghost file' in cloudinary, with no corresponding DB document.
To fix this, we will be tagging the file on the intial upload with the
presigned URL, and upon the backends 2nd request, we will be untagging
this tag. Then we will create a cron job that periodically scans our
cloudinary, and removes any files with this special tag. I guess the
only problem with this would be that when we run this job, we could
delete the files of a user that is in the middle of their upload request
cycle (they are waiting for backend to remove tag). But we could easily
overcome this by checking something like tag existance AND the file is
older than 24 hours