Added 'visibility'. Slight changed errors & routes
-products can now have .visibility of public or private
-errors now use error.name instead of error.type, because external api
calls and normal errors come back with a .name
-admin routes now are preceded by 'admin', like 'admin/products', while
public routes are just products

TODO
-Create a cron job that periodically scans our cloudinary, and removes
all files with context value of 'linked= false', that are also older
than 24 hours. This prevents 'ghost files', because after the client
gets our presigned URL from 1st backend request, and then uses this to
uploaded directly from frontend to cloudinary, we set 'linked=false'.
Then when their upload is successful, the client requests our backend
again, and this is when we then set it to true. But for example, if for
some reason this request to our backend isn't fired, then we'd end up
with an uploaded file with no DB document referencing it.
The reason we need it to ALSO be older than 24 hours, is so that we
don't delete files that are in the middle of this same upload process,
that a user tried to upload 1s ago, but is being mistaken as a 'ghost
file' because in the middle of these requests, it has 'linked=false'