DB Schema supports multiple images. Routes work
Deleting Product also work

TODO
-In the Product Schema, add an 'order' field to each image, also expect
this property from the route input 'images' as in input:
{images: [{publicId, order}]}
-actually support uploading multiple images, and editing in multiple
images in a product on Edit. Delete product works well already
-Add currently has a 'rollback' that if DB operation goes wrong, then
delete image. Implement this same idea when switching it to multiple
images. For edit product not sure if I would do the same, have to think
about this.
Originally my solution is to use a trasaction in '/edit' like:
1st= edit DB documents to exclude images that were edited out
2nd= Delete from cloudinary. If errors here, then the trasaction would
rollback the document
PREV TODO
-Create a cron job that periodically scans our cloudinary, and removes
all files with context value of 'linked= false', that are also older
than 24 hours. This prevents 'ghost files', because after the client
gets our presigned URL from 1st backend request, and then uses this to
uploaded directly from frontend to cloudinary, we set 'linked=false'.
Then when their upload is successful, the client requests our backend
again, and this is when we then set it to true. But for example, if for
some reason this request to our backend isn't fired, then we'd end up
with an uploaded file with no DB document referencing it.
The reason we need it to ALSO be older than 24 hours, is so that we
don't delete files that are in the middle of this same upload process,
that a user tried to upload 1s ago, but is being mistaken as a 'ghost
file' because in the middle of these requests, it has 'linked=false'