Refactored addProduct, next test it
Also small change to validation, now we are attaching the parsed
validated data into req.validatedData, instead of not putting it
anywhere. I needed to attach it somewhere like this, because we
are applying a transformation in our zod validation schema that
modofies the validated data. So now we can access this transformed data

TODO
Will add user auth now
refer to prev commit