Added Google Auth with MongoStore for sessions
Improved a lot of different small things on the auth
Added middleware that modifies res.json() so that if req.user exists,
then calling res.json will always add a field of 'user' with the
req.user from passport. This way, when we call res.json on every route
and on the catch all error handler, it will always attach the user
object if they are signed in to a session currently

Also improved dev vs production by:
- adding library "cross-env", then setting package.json scripts to set
env variables like in unix
- specifying the redirect page on auth/google/redirect to redirect to
a differnt page deppending on whether NODE_ENV is in dev or production

TODO
-add route to let any user change their user level between admin and
user (not owner)
-deploy anywhere to have as a demo
-add cart Database function + route
-refer to prev commits, I think it was 2-3 commits ago when working on
addProduct (small things I think). And then need big refactor of
editProduct I think