Added mandatory auths + user level checks. Deploy
Also separated getProducts into a public route and an admin route. The
reason for this is that in the frontend, when I am fetching the admin
panel, then I want to be able to get an error that says I do not have
admin powers. Separating it into 2 routes makes this easy. Also the
code in the backend for this is quite short & simple.

TODO
now deploy
refer to many prev commits