Implemented editProducts, to test. Moved category
Moved categories to its own controller + added userLevel checks
between the req.user and the creator of the category.
Also extracted the function that marks images as "linked=true" when
a user tries to save a product info with its images, so it can be
used in both addProduct and editProduct. Maybe move this to a
/services/cloudinary.js

TODO
test edit products and categoirs and everything. Clean up + remove some
comments
refer to many prev commits