Adding websockets - bootleg. Fixed google sign-in
There was an issue where if you only had 1 google account saved on
your browser, and clicked the sign-in button, then our passport google
auth would instantly sign-them in without bringing up the google
sign-in page where you can choose to sign-in with a different account.
Fixed by a simple `prompt: 'select_account'` in a options in auth.js
passport.authenticate

In the middle of addding websockets
Purpose: The purpose is to replace our logic that sends a .user object
attached to every request (a lot of extra network data), where we
currently use every single request, inspect its .user field, and
overwrite our redux user state...
And instead replace above with 1 user fetch on app load + user session
updates via websockets
Progress:
- Added socket.io to our express server: merged the existing express
server and websockets, and merged our passport sessions, so that
'io' has access to all this information
- simple barebones connection flow works to connect clients signed in
to a user session, to the same room as anyone with the same session
(basically connects same browser tabs to the same websocket room)

TODO
- refer to prev commits
- there are still many things to address to integrate webscockets,
mostly comented. For ex: need to chaneg one of the middlewares
that runs when handshaking, so that it tells the frontend to stop
polling endlessly if you are not signed in
Also need to figure out a way to disconnect users when their cookie
expires. I think if our app was really time important, there are a
many diff options, and I still need to investigate further all
possible solutions, but one solution is maybe on the frontend if I get
a 401, then emit a websocket event that checks my user session status !