Working websockets to update user data changes
Fully works to update user state accross browser tabs, when one of those
tabs signs AND/OR logs out.

TODO
- refer to prev commits
- a lot of cleaning up
(write the the following somwhere in a comment or notes.txt or README):
We initially just used 1 websocket, and had to choose between joining
the client to a room based on their user.id or passport session.
I then realized though that I would need to join them to both rooms,
because I want to send the same events per single client: right now
for logging in & out. I also wanted to join them to their user.id room
to emit events in teh future for user updates.
However, sessions in passport change everytime you log in and out, and
additionally in passport.authenticate('google'), there are specific
sports where this data becomes innaccessible where it needs to be.
So then I solved those problems and did some back and forht with
joining and leaving different rooms w/ this.
And this almost worked, and in the scenario where 1 browser opened 2
tabs while signed in, both tabs shared the same sessionId, however
if there wasnt a signed in user, then they did NOT share the same
session id. This is due to sessions having a setting 'saveUnitialized',
which determines if non-signed in tabs will be saved to your store or
not. We had it set to false, which means it only creates DB documents
when a user signs in, but this meant that 2 tabs on the same browser
will always create 2 diff sessions....
The problem is that when I set it to false, while it solved our
socket room issue, it also created a new big scaling probelm where
everytime a new tab would load our site, it would always write a new
DB document for a new session.... this is very bad scaling-wise.....

So then I pivoted to our final solution:
When a browser visits our page, and does not have a cookie 'clientId',
then we will create one for them. We then use this clientId as the
room id to join tabs on the same browser to the same room.
I also had to do a bit of req/cookies gymnastics because of some
problems with session clearing the cookies and clearing the req body
in between certain steps in the passport authenticate following