Updated username creation to use 1st name + nanoid
TODO
Important.
This is all after I have migrated prject-gg from this same current
auth method, to now use BetterAuth

TODO
We learned there a couple of things:
- Its best to not embed authorized accounts in the way that we do in
Users here, and instead create a spearate 'accounts' table where each
document is a new type of account 'google, github, etc', and each doc
has a .user field pointing to the user the account belongs to
- Subscriptions schema should changr. I know we have a UserLevel schema,
but we should proabbly follow the same approach. Although in project-gg
it was hard to decide on its schema, Its best to mode each subscription
like {id, tier, name} , where id and tier are the same number. This way
anything that interacts with User can decide whether it wants to call
populate() on it to get the name and extra Subsciption info, otherwise
it can just NOT populate(), and read the user.subscription directly,
since that id will be a number like 1, 2 ,3 and this id will be the
same number as the value of the 'tier' field